# remote/swagger.yaml
basePath: /API
consumes:
    - application/json
definitions:
    ErrorResponse:
        description: 'Error response.'
        properties:
            responseCode:
                type: number
                format: double
                description: 'error code in system.'
            message:
                type: string
                description: 'Readble message.'
                x-nullable: true
        required:
            - responseCode
        type: object
    Login:
        description: 'Login schema model.'
        properties:
            email:
                type: string
            password:
                type: string
        required:
            - email
            - password
        type: object
...
    LocalServerInfo:
        description: 'Info about local server.'
        properties:
            localServerId:
                type: string
                description: 'Local server id'
            displayName:
                type: string
                description: 'Display name'
        required:
            - localServerId
            - displayName
        type: object
    LoginLocalServer:
        description: "Extends login with local server selection, case use own more then one local server.\r\nneeds to know witch server to connect."
        properties:
            email:
                type: string
            password:
                type: string
            localServerId:
                type: string
                description: "Local server to login to. \r\nif user exsit only in one local server ignore field."
                x-nullable: true
        required:
            - email
            - password
        type: object
...
info:
    title: 'CASAnet - remote server API'
    version: 2.0.0
    description: 'Remote server for forwarding requests to locals home network. hosted in [GitHub repo](https://github.com/haimkastner/Home-IoT-Server).'
    license:
        name: ISC
paths:
    /auth/login:
        post:
            operationId: LoginDocumentation
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: object
                '201':
                    description: '2-fatore code sent'
                '210':
                    description: 'select local server to connect to'
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/LocalServerInfo'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: "Login to local server via remote server channel.\r\nIf users exsits in more then one local server, it`s return status code 210 and the avalible user servers to select.\r\n     *"
            tags:
                - Authentication
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/LoginLocalServer'
    /auth/login/tfa:
        post:
            operationId: LoginTfaDocumentation
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: '2-step verification login to local server via remote server channel.'
            tags:
                - Authentication
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/LoginLocalServer'
    /auth/logout:
        post:
            operationId: LogoutDocumentation
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: 'Logout manualy from remote and local server systems.'
            tags:
                - Authentication
            security:
                -
                    userAuth: []
                -
                    adminAuth: []
            parameters: []
...
    /administration/auth/login:
        post:
            operationId: AdministrationLoginDocumentation
            produces:
                - application/json
            responses:
                '201':
                    description: '2-fatore code sent'
                '204':
                    description: 'No content'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: 'Login to remote administration system.'
            tags:
                - Administration
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/Login'
    /administration/auth/login/tfa:
        post:
            operationId: AdministrationLoginTfaDocumentation
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '403':
                    description: 'Auth fail'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '422':
                    description: 'Invalid schema'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: '2-step verification login to remote administration.'
            tags:
                - Administration
            security: []
            parameters:
                -
                    in: body
                    name: login
                    required: true
                    schema:
                        $ref: '#/definitions/Login'
    /administration/auth/logout:
        post:
            operationId: AdministrationLogoutDocumentation
            produces:
                - application/json
            responses:
                '204':
                    description: 'No content'
                '501':
                    description: 'Server error'
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            description: 'Logout manualy from administration system.'
            tags:
                - Administration
            security:
                -
                    adminAuth: []
            parameters: []
...
produces:
    - application/json
swagger: '2.0'
securityDefinitions:
    userAuth:
        type: apiKey
        name: Cookie
        in: header
    adminAuth:
        type: apiKey
        name: Cookie
        in: header
host: 'localhost:3000'
...
schemes:
    - http
    - https